scripts/salam_gui/requirements.txt
# gem5-SALAM GUI Dependencies
PySide6>=6.2.0
pyqtgraph>=0.13.0
networkx>=3.0
pydot>=1.4.0
pandas>=2.0.0
matplotlib>=3.7.0
PyYAML>=6.0
pyzmq>=25.0.0
